feature/ERP: Create Utilities for Email
